Added jsonResponse. Refactored code and removed unused dependecy for data attribute.
