+++ go.mod
@@ -6,5 +6,4 @@ require (
 	github.com/davecgh/go-spew v1.1.1
 	github.com/gorilla/mux v1.8.0
 	github.com/joho/godotenv v1.3.0
-	github.com/shopspring/decimal v1.2.0
 )
